update #14: reference resolution and pathing
Lots of cool stuff here.  All the $ref and $dynamicRef functions are
complete.  The schema path tracking functions are complete.  Within the
$ref resolution, schema can now be sourced from .txt files via the $ref
keyword, and the retrieved schema will be validated and saved to the
database for future reference.  The schema_validate_Format() function
has also been fleshed out so that all json-schema.org defined formats
are validated.