update #49: another quite large update; almost all testing is complete and most functionality is in place
This commit contains a lot of modified functions and function names.
There were a lot of changes while running various json-schema.org test
suites to ensure full compliance with specification.  Fully annonations
as described in the spec are built in and used, especially for
unevaluated* keywords.  Too much more to describe here.